[merb-assets] Adding possibility to set charset for js_include_tag.
Adding default charset attribute to the js_include_tag in the same way
as css_include_tag works right now. So default charset is utf-8 and is
it possible to override it using :charset => 'xy' option.

Fixing specs to pass on 1.9.1.
The issue would probably be valid for other versions and implementations
as well. Since it involves order how attributes are ordered inside the
script tag.

I think we should check for the attribute existence and
value not the order.